Make ICorsOptions fields optional
None of these are required but the type doesn't reflect that and forces users consuming the API via TypeScript to either cast their type to any or fill in the default values manually.

Looking at other middleware options it would seem they all follow the same pattern of being optional.