feat/types/remove-aws-xray-sdk-deps
In several plugin definitions `aws-xray-sdk` was used. The problem is that we need to be able to continue using those plugins without being dependent on `aws-xray-sdk`.

In this PR I propose to remove this dependency, and to use the `Options` interface defined in `utils` to define the interfaces of each dependent plugins.
This allows a simpler management of types, in case of update of `aws-xray-sdk`, we will have only one element to modify. The type of `awsClientCapture` in `utils.Options`.

There are some plugins that use only one part of the attributes of the "Option" interface, for this case I use the typescript utility type : [`Pick`](https://www.typescriptlang.org/docs/handbook/utility-types.html#picktype-keys)